Console commands highlighted as such
These commands are not from a shell script, but should be ran in a shell
session. `console` is the highlighter that works for GitHub. I also
added truncated output of the `git clone` command to make it clear when
it’s time to `cd`. In the future, perhaps adding other truncated
sessions would help add context to said commands.